+++ Basic/Count-and-Length/PSCustomObject/Test-1.normal.vs.custom.ps1
@@ -13,6 +13,7 @@ Set-StrictMode -Off
 # "normal" objects have surrogate Count
 "normal.Count: $($normal.Count)"
 
+# v6.1.0 -- fixed
 # "custom" objects do not have surrogate Count
 # $custom.Count is either $null or error in the strict mode
 "custom.Count: $($custom.Count)"