+++ Basic/Count-and-Length/PSCustomObject/Test-2.Select-Object.ps1
@@ -15,6 +15,7 @@ Set-StrictMode -Off
 "Get-Some 2 : $((Get-Some 2).Count)"
 "Get-Some 1 : $((Get-Some 1).Count)"
 
+# v6.1.0 -- fixed, gets expected 1
 # get Count 2 (expected) and $null/error (may be unexpected)
 "Get-Some 2 | Select-Object : $((Get-Some 2 | Select-Object Name).Count)"
 "Get-Some 1 | Select-Object : $((Get-Some 1 | Select-Object Name).Count)"