+++ Basic/Number-of-returned-objects/test5.PSCustomObject.ps1
@@ -1,7 +1,10 @@
 
-$Version = $PSVersionTable.PSVersion.Major
-if ($Version -ge 3) {
-	Set-StrictMode -Off
+$Version = $PSVersionTable.PSVersion
+if ($Version -ge ([version]'3.0.0') -and $Version -lt ([version]'6.0.9999')) {
+    Set-StrictMode -Off
+}
+else {
+    Set-StrictMode -Version Latest
 }
 
 $object = [PSCustomObject]@{Name='Joe'; Age=42}
@@ -10,6 +13,9 @@ $object = [PSCustomObject]@{Name='Joe'; Age=42}
 # v3+:
 # - $null, $true, PSCustomObject
 # - error in strict mode (v5)
+# v6.1.0
+# - 1, $false, PSCustomObject
+# - no error in strict mode
 
 $object.Count
 $null -eq $object.Count