+++ Cmdlets/Get-ChildItem/v5-LiteralPath-Recurse-ignores-Include/.test.ps1
@@ -1,5 +1,6 @@
 
 $Version = $PSVersionTable.PSVersion.Major
+$v610 = $PSVersionTable.PSVersion -ge ([version]'6.0.9999')
 
 task Test-1.Path {
 	$r = .\Test-1.Path.ps1
@@ -11,7 +12,7 @@ task Test-2.Path {
 	$r = .\Test-2.LiteralPath.ps1
 	($r = @($r | Group-Object Extension -NoElement))
 
-	if ($Version -ge 5 -or $Version -eq 2) {
+	if ($Version -eq 2 -or ($Version -ge 5 -and !$v610)) {
 		# unexpected
 		assert ($r.Count -ge 2)
 	}